requirements.txt
asgiref==3.7.2
distro==1.8.0
Django==4.2.5
django-sesame==3.2.1
et-xmlfile==1.1.0
JPype1==1.4.1
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
psycopg2==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tabula-py==2.8.1
typing_extensions==4.7.1
tzdata==2023.3
